Implement a random data generator, and use it to test the decoder.
The random data generator should include all cases that can occur in reality,
including scenarios that aren't generated by open-vcdiff (such as using a
substring of the dictionary as the source segment), and some edge cases that
probably shouldn't occur in practice (like instructions with zero output size).

Besides the unit test, I added the ability to generate and store the source,
delta and target files as files. This is useful for verifying that the decoded
target data also matches the output of the open-vcdiff decoder (which it does).